refactor(003): relocate PathEnvironmentVariable to CliInvoke; drop FilePathResolverBase
Relocate PathEnvironmentVariable from CliInvoke.Core into CliInvoke as an
internal static class, and remove the public abstract FilePathResolverBase
from Core (v3 breaking, D006). FilePathResolver now implements IFilePathResolver
directly and uses the CliInvoke-side PathEnvironmentVariable; PATH-first then
directory-recursion order and the Try* catch(Exception) discipline are
preserved. The test stub now derives from FilePathResolver instead of the
removed base class.
