+++ src/CliInvoke/Internal/IO/PathEnvironmentVariable.cs
@@ -1,18 +1,18 @@
 /*
     MIT License
-   
+
     Copyright (c) 2025-2026 Alastair Lundy
-   
+
     Permission is hereby granted, free of charge, to any person obtaining a copy
     of this software and associated documentation files (the "Software"), to deal
     in the Software without restriction, including without limitation the rights
     to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
     copies of the Software, and to permit persons to whom the Software is
     furnished to do so, subject to the following conditions:
-   
+
     The above copyright notice and this permission notice shall be included in all
     copies or substantial portions of the Software.
-   
+
     THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
     IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
     FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -20,13 +20,13 @@
     LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
     OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
     SOFTWARE.
- */
+*/
 
 // Derived from DotExtensions (MIT License) - See THIRD_PARTY_NOTICES.txt for full details.
 
 using System.Linq;
 
-namespace CliInvoke.Core.Internal.IO;
+namespace CliInvoke.Internal.IO;
 
 /// <summary>
 /// Provides utility methods for interacting with the system's PATH environment variable
@@ -38,7 +38,7 @@ internal static class PathEnvironmentVariable
     /// Represents the character used to separate individual entries in the PATH environment variable.
     /// The value is ';' on Windows and ':' on non-Windows operating systems.
     /// </summary>
-    public static char PathContentsSeparatorChar 
+    public static char PathContentsSeparatorChar
         => OperatingSystem.IsWindows() ? ';' : ':';
 
     /// <summary>
@@ -77,11 +77,11 @@ internal static class PathEnvironmentVariable
                     return
                         $"{x.Substring(0, homeTokenIndex)}{userProfile}{x[(homeTokenIndex + homeToken.Length)..]}";
                 }
-                
+
                 return x.TrimEnd(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar);
             });
     }
-    
+
     /// <summary>
     /// Retrieves the directories listed in the system's PATH environment variable.
     /// </summary>
@@ -103,7 +103,7 @@ internal static class PathEnvironmentVariable
     internal static IEnumerable<string> EnumerateFileExtensions()
     {
         if (!OperatingSystem.IsWindows()) return [""];
-        
+
         return Environment
                    .GetEnvironmentVariable("PATHEXT")
                    ?.Split(PathContentsSeparatorChar, StringSplitOptions.RemoveEmptyEntries)
@@ -112,7 +112,7 @@ internal static class PathEnvironmentVariable
                    {
                        x = x.Trim();
                        x = x.Trim('"');
-                       
+
                        if (!x.StartsWith('.', StringComparison.OrdinalIgnoreCase))
                            x = x.Insert(0, ".");
